refactor: add useSearch hook
