+++ src/app/phrases/types/phrases.d.ts
@@ -2,3 +2,5 @@ export interface Phrase {
   id: string;
   content: string;
 }
+
+export type OnSearchChange = (searchTerm: string) => void;
